Renamed sph_kernel_tables.hpp → sph_kernels.hpp
