+++ sph/include/sph/hydro_std/iad_kern.hpp
@@ -5,7 +5,7 @@
 #include "cstone/traversal/ijloop/ijloop.hpp"
 
 #include "sph/iad_regularization.hpp"
-#include "sph/sph_kernel_tables.hpp"
+#include "sph/sph_kernels.hpp"
 
 namespace sph
 {
