+++ sph/include/sph/hydro_ve/divv_curlv_kern.hpp
@@ -33,7 +33,7 @@
 
 #include "cstone/traversal/ijloop/ijloop.hpp"
 
-#include "sph/sph_kernel_tables.hpp"
+#include "sph/sph_kernels.hpp"
 
 namespace sph
 {
