+++ sph/include/sph/hydro_ve/iad_gradh_kern.hpp
@@ -36,7 +36,7 @@
 #include "cstone/traversal/ijloop/ijloop.hpp"
 
 #include "sph/iad_regularization.hpp"
-#include "sph/sph_kernel_tables.hpp"
+#include "sph/sph_kernels.hpp"
 
 namespace sph
 {
